sine_wave/requirements.txt
action-msgs==2.0.2
action-tutorials-interfaces==0.33.5
action-tutorials-py==0.33.5
actionlib-msgs==5.3.5
ament-cmake-test==2.5.3
ament-copyright==0.17.1
ament-cppcheck==0.17.1
ament-cpplint==0.17.1
ament-flake8==0.17.1
ament-index-python==1.8.1
ament-lint==0.17.1
ament-lint-cmake==0.17.1
ament-package==0.16.3
ament-pep257==0.17.1
ament-uncrustify==0.17.1
ament-xmllint==0.17.1
angles==1.16.0
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrdict==2.0.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
builtin-interfaces==2.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clarabel==0.9.0
click==8.1.7
cloudpickle==3.1.0
cmake-generate-parameter-module-example==0.4.0
colorama==0.4.6
comm==0.2.2
composition-interfaces==2.0.2
ConfigSpace==1.2.0
contourpy==1.3.1
control-msgs==5.3.0
controller-manager==4.24.0
controller-manager-msgs==4.24.0
copier==9.2.0
cv-bridge==4.1.0
cvxpy==1.6.0
cycler==0.12.1
dask==2024.11.1
dask-jobqueue==0.9.0
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
demo-nodes-py==0.33.5
diagnostic-msgs==5.3.5
diagnostic-updater==4.2.1
distributed==2024.11.1
domain-coordinator==0.12.0
dunamai==1.23.0
emcee==3.1.6
example-interfaces==0.12.0
examples-rclpy-executors==0.19.4
examples-rclpy-minimal-action-client==0.19.4
examples-rclpy-minimal-action-server==0.19.4
examples-rclpy-minimal-client==0.19.4
examples-rclpy-minimal-publisher==0.19.4
examples-rclpy-minimal-service==0.19.4
examples-rclpy-minimal-subscriber==0.19.4
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.10.0
funcy==2.0
generate-parameter-library-py==0.4.0
generate-parameter-module-example==0.4.0
geometry-msgs==5.3.5
grayscale-image==0.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
image-geometry==4.1.0
imageio==2.36.0
iniconfig==2.0.0
interactive-markers==2.5.4
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jinja2-ansible-filters==1.3.2
jinja2-strcase==0.0.2
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
laser-geometry==2.7.0
launch==3.4.3
launch-ros==0.26.6
launch-testing==3.4.3
launch-testing-ros==0.26.6
launch-xml==3.4.3
launch-yaml==3.4.3
lazy_loader==0.4
lifecycle-msgs==2.0.2
locket==1.0.0
logging-demo==0.33.5
map-msgs==2.4.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
message-filters==4.11.3
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
nav-msgs==5.3.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
osqp==0.6.7.post3
osrf-pycommon==2.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pcl-msgs==1.0.0
pendulum-msgs==0.33.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
plumbum==1.9.0
progressbar==2.5
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
Pympler==1.1
pynisher==1.0.10
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyquaternion==0.9.9
pyrfr==0.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-qt-binding==2.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post4
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.5
questionary==2.1.0
rcl-interfaces==2.0.2
rclpy==7.1.3
rcutils==6.7.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
resource-retriever==3.4.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rmw-dds-common==3.1.0
ros-image-to-qimage==0.4.1
ros2-pkg-create==1.1.0
ros2action==0.32.2
ros2bag==0.26.6
ros2bag-mcap-cli==0.26.6
ros2bag-sqlite3-cli==0.26.6
ros2cli==0.32.2
ros2component==0.32.2
ros2doctor==0.32.2
ros2interface==0.32.2
ros2launch==0.26.6
ros2lifecycle==0.32.2
ros2multicast==0.32.2
ros2node==0.32.2
ros2param==0.32.2
ros2pkg==0.32.2
ros2run==0.32.2
ros2service==0.32.2
ros2topic==0.32.2
rosbag2-interfaces==0.26.6
rosbag2-py==0.26.6
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.5
rosidl-cli==4.6.5
rosidl-cmake==4.6.5
rosidl-generator-c==4.6.5
rosidl-generator-cpp==4.6.5
rosidl-generator-py==0.22.0
rosidl-generator-type-description==4.6.5
rosidl-parser==4.6.5
rosidl-pycommon==4.6.5
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.1
rosidl-typesupport-fastrtps-cpp==3.6.1
rosidl-typesupport-introspection-c==4.6.5
rosidl-typesupport-introspection-cpp==4.6.5
rpds-py==0.21.0
rpyutils==0.4.1
rqt==1.6.0
rqt-action==2.2.0
rqt-bag==1.5.4
rqt-bag-plugins==1.5.4
rqt-console==2.2.1
rqt-controller-manager==4.24.0
rqt-dotgraph==0.0.4
rqt-gauges==0.0.3
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-joint-trajectory-controller==4.19.0
rqt-moveit==1.0.1
rqt-msg==1.5.1
rqt-plot==1.4.0
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.6
rqt-robot-steering==1.0.0
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-tf-tree==1.0.5
rqt-topic==1.7.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
scs==3.2.7
Send2Trash==1.8.3
sensor-msgs==5.3.5
sensor-msgs-py==5.3.5
service-msgs==2.0.2
setuptools==75.4.0
shape-msgs==5.3.5
sine-wave==0.0.0
six==1.16.0
smac==2.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sros2==0.13.2
stack-data==0.6.3
statistics-msgs==2.0.2
std-msgs==5.3.5
std-srvs==5.3.5
stereo-msgs==5.3.5
sympy==1.13.1
tblib==3.0.0
teleop-twist-keyboard==2.4.0
terminado==0.18.1
tf2-geometry-msgs==0.36.8
tf2-kdl==0.36.8
tf2-msgs==0.36.8
tf2-py==0.36.8
tf2-ros-py==0.36.8
tf2-sensor-msgs==0.36.8
tf2-tools==0.36.8
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
toolz==1.0.0
topic-monitor==0.33.5
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
traitlets==5.14.3
trajectory-msgs==5.3.5
triton==3.1.0
turtlesim==1.8.3
type-description-interfaces==2.0.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
unique-identifier-msgs==2.5.0
uri-template==1.3.0
urllib3==2.2.3
visualization-msgs==5.3.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.0
widgetsnbextension==4.0.13
zict==3.0.0
